requirements.txt
crewai-tools==0.40.1
google-generativeai==0.8.4
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3